se modifica la captura de errores

+++ domain/model/src/main/java/co/com/crediya_solicitud/model/loantypes/LoanTypes.java
@@ -15,7 +15,7 @@ public class LoanTypes {
     private String name;
     private BigDecimal minimumAmount;
     private BigDecimal maximumAmount;
-    private Double interestRate;
+    private Integer interestRate;
     private Boolean automaticValidation;
 
     public boolean isValidAmount(BigDecimal amount) {
